apply similar ideas to the linear and circular lattice q-vectors and then make them all consistent so they use the same searching method for the lattice point search
